Internet Option Function
And many other things

+++ dash/src/Template/Students/edit.ctp
@@ -1,23 +1,49 @@
-<div class="actions columns large-2 medium-3">
-    <h3><?= __('Menu') ?></h3>
-    <ul class="side-nav">
-        <li><?= $this->Form->postLink(__('Delete This Student'), ['action' => 'delete', $student->id], ['confirm' => __('Are you sure you want to delete # {0}?', $student->id)]) ?> </li>
-        <li><a href="javascript:history.back()">Go Back</a></li>
-    </ul>
-</div>
-<div class="students form large-10 medium-9 columns">
-    <?= $this->Form->create($student); ?>
-    <fieldset>
-        <legend><?= __('Edit Student') ?></legend>
-        <?php
-            echo $this->Form->input('person_id', ['options' => $person]);
+<?php $currentlogged = $this->Session->read('Auth.User'); ?>
+
+<?php if ($currentlogged['role'] === "admin") : ?>
+
+    <div class="actions columns large-2 medium-3">
+        <h3><?= __('Menu') ?></h3>
+        <ul class="side-nav">
+            <li><?= $this->Form->postLink(__('Delete This Student'), ['action' => 'delete', $student->id], ['confirm' => __('Are you sure you want to delete # {0}?', $student->id)]) ?> </li>
+            <li><a href="javascript:history.back()">Go Back</a></li>
+        </ul>
+    </div>
+    <div class="students form large-10 medium-9 columns">
+        <?= $this->Form->create($student); ?>
+        <fieldset>
+            <legend><?= __('Edit Student') ?></legend>
+            <?php
+                echo $this->Form->input('person_id', ['options' => $person]);
+                echo $this->Form->input('internet_plan', ['options' => ['NONE' => 'NONE', 'BASIC' => 'BASIC', 'STANDARD' => 'STANDARD', 'PREMIUM' => 'PREMIUM']]);
+            ?>
+        </fieldset>
+        <?= $this->Form->button(__('Submit')) ?>
+        <?= $this->Form->end() ?>
+        <?= $this->Form->create(null, [
+            'url' => ['controller' => 'Students', 'action' => 'index']
+        ])?>
+        <?= $this->Form->button(__('Cancel')) ?>
+    </div>
+
+<?php endif; ?>
+
+<?php if ($currentlogged['role'] === "tenant") : ?>
+
+    <div class="students form large-10 medium-9 columns">
+        <?= $this->Form->create($student); ?>
+        <fieldset>
+            <legend><?= __('Edit Student') ?></legend>
+            <?php
             echo $this->Form->input('internet_plan', ['options' => ['NONE' => 'NONE', 'BASIC' => 'BASIC', 'STANDARD' => 'STANDARD', 'PREMIUM' => 'PREMIUM']]);
-        ?>
-    </fieldset>
-    <?= $this->Form->button(__('Submit')) ?>
-    <?= $this->Form->end() ?>
-    <?= $this->Form->create(null, [
-        'url' => ['controller' => 'Students', 'action' => 'index']
-    ])?>
-    <?= $this->Form->button(__('Cancel')) ?>
-</div>
+            ?>
+        </fieldset>
+        <?= $this->Form->button(__('Submit')) ?>
+        <?= $this->Form->end() ?>
+        <?= $this->Form->create(null, [
+            'url' => ['controller' => 'Students', 'action' => 'index']
+        ])?>
+        <?= $this->Form->button(__('Cancel')) ?>
+    </div>
+
+<?php endif; ?>
